Wallet Connection: Use the "Wild Tangz" Connector
This is the wallet connector found in:
https://github.com/thaddeusdiamond/cardano-nft-js-toolkit

With a little bit of manipulation so that it can fit inside of the React
framework.

The `cardano-nft-js-toolkit` package is intended for usage in plain old
vanilla JavaScript (e.g., Squarespace).  We attempted to bundle it, but
it does not currently play nice with npm/react, so we abandoned that
attempt.

Instead, we create a static link to the hosted `wildtangz.pages.dev`
version of the wallet connector, which binds the `NftToolkit` object to
the `window` object.

Then, inside of our component, we can add an effect to process the
NftToolkit components on page load.  Once we do that, we can access the
cardanoDApp as normal when a user connects using functional closures.

If we later want to work with `cardanoDapp` we can either retrieve it
from the `NftToolkit` object bound to `window`, or we can pull state
from the local JSX component.

NOTE: Styling is left to the reader (though toastify.js is included so
it does not mess with the DOM too much).

[ Documentation: Code itself ]
[ Testing: npm run develop ]